requirements.txt
# Runtime
scrapy==2.11.1
scrapy-playwright==0.0.33
playwright==1.47.0
pydantic==2.9.2
PyYAML==6.0.1
requests==2.32.3
orjson==3.10.3
httpx==0.27.2
shapely==2.0.5
geopandas==1.0.1
asyncpg>=0.28.0
psycopg2-binary>=2.9.7
fastapi
schedule

# Dev & tests
pytest==8.3.2
